Added address registration
Added very basic address registration. Need to update this and the database for address as number and street name. Then create a query to input all the information into the database. Found an issue where if you login when the account doesn't exist, it causes a crash trying to check the password.