Adding images
Began to add images. We can now add images in the HTML which is pushed back to Flask. A function to convert the image has been written, but is still unimplemented for the moment. The commit to the DB will fail until we pass the path of the image, instead of the image itself